Refactor app.js
